FuncionSpark/function/requirements.txt
functions-framework==3.*
google-cloud-dataproc==1.0.0